requirements.txt
# Core dependencies
fastapi==0.95.2
uvicorn==0.22.0

# Database
sqlite3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.3.1
pytest-cov==4.0.0
requests==2.28.2

# Optional: For development
python-dotenv==1.0.0